Refactor PDF generation into modular pdf module
Replaces the monolithic pdf_generator.rs with a new modular src/pdf/ structure, splitting PDF generation logic into commands, common, cotizacion, and informe modules. Updates references in lib.rs to use the new pdf module and removes the old pdf_generator.rs file. This improves maintainability and separation of concerns for PDF-related code.
