+++ src-tauri/src/pdf/mod.rs
@@ -1,0 +1,11 @@
+pub mod common;
+pub mod cotizacion;
+pub mod informe;
+pub mod commands;
+
+pub use common::*;
+pub use cotizacion::{CotizacionPdfData, CotizacionPdfGenerator};
+pub use informe::{InformePdfData, InformePdfGenerator};
+pub use commands::{generate_cotizacion_pdf_command, generate_informe_pdf_command};
+
+
